+++ packages/desktop/src/renderer/components/Channel/Channel.stories.cy.tsx
@@ -11,6 +11,7 @@ import { HTML5Backend } from 'react-dnd-html5-backend'
 import { DisplayableMessage } from '@quiet/types'
 
 import ChannelComponent from './ChannelComponent'
+import { payloadDuplicated, payloadUnregistered } from '../widgets/userLabel/UserLabel.types'
 
 const Template: ComponentStory<typeof ChannelComponent> = () => {
   const [messages, setMessages] = useState<{
@@ -28,7 +29,7 @@ const Template: ComponentStory<typeof ChannelComponent> = () => {
       message: message,
       createdAt: 0,
       date: '12:46',
-      nickname: 'vader'
+      nickname: 'vader',
     }
     const _messages = mock_messages(_message)
     setMessages(_messages)
@@ -48,23 +49,23 @@ const Template: ComponentStory<typeof ChannelComponent> = () => {
             createdAt: 0,
             channelId: 'general',
             signature: 'signature',
-            pubKey: 'pubKey'
+            pubKey: 'pubKey',
           }}
           user={{
             id: 'id',
             nickname: 'vader',
             hiddenService: {
               onionAddress: 'onionAddress',
-              privateKey: 'privateKey'
+              privateKey: 'privateKey',
             },
             peerId: {
               id: 'id',
               privKey: 'privKey',
-              pubKey: 'pubKey'
+              pubKey: 'pubKey',
             },
             dmKeys: {
               publicKey: 'publicKey',
-              privateKey: 'privateKey'
+              privateKey: 'privateKey',
             },
             userCsr: {
               userCsr: 'userCsr',
@@ -72,18 +73,18 @@ const Template: ComponentStory<typeof ChannelComponent> = () => {
               pkcs10: {
                 publicKey: 'publicKey',
                 privateKey: 'privateKey',
-                pkcs10: 'pkcs10'
-              }
+                pkcs10: 'pkcs10',
+              },
             },
             userCertificate: 'userCertificate',
-            joinTimestamp: null
+            joinTimestamp: null,
           }}
           isCommunityInitialized={true}
           uploadedFileModal={{
             open: false,
             handleOpen: function (_args?: any): any {},
             handleClose: function (): any {},
-            src: 'images/butterfly.jpeg'
+            src: 'images/butterfly.jpeg',
           }}
           channelId={'general'}
           channelName={'general'}
@@ -107,6 +108,8 @@ const Template: ComponentStory<typeof ChannelComponent> = () => {
           handleClipboardFiles={function (arg: ArrayBuffer, ext: string, name: string): void {
             throw new Error('Function not implemented.')
           }}
+          duplicatedUsernameModalHandleOpen={() => payloadDuplicated}
+          unregisteredUsernameModalHandleOpen={() => payloadUnregistered}
         />
       </DndProvider>
     </>
@@ -118,7 +121,7 @@ export const Component = Template.bind({})
 const component: ComponentMeta<typeof ChannelComponent> = {
   title: 'Components/ChannelComponentCypress',
   decorators: [withTheme],
-  component: ChannelComponent
+  component: ChannelComponent,
 }
 
 export default component